10 - implement FileHelper class. Added exceptions to FileHelper.class, prepared FileHelperTest.class

+++ src/main/java/pl/coderstrust/repository/FileHelper.java
@@ -13,6 +13,9 @@ public class FileHelper {
   private File file = null;
 
   public void create(String filePath) throws IOException {
+    if (filePath == null){
+      throw new IllegalArgumentException("filePath cannot be null.");
+    }
     if (file == null) {
       this.file = new File(filePath);
       file.createNewFile();
@@ -20,6 +23,9 @@ public class FileHelper {
   }
 
   public void delete(String filePath) throws IOException {
+    if (filePath == null){
+      throw new IllegalArgumentException("filePath cannot be null.");
+    }
     create(filePath);
     file.delete();
   }
@@ -33,17 +39,29 @@ public class FileHelper {
   }
 
   public void clear(String filePath) throws FileNotFoundException {
+    if (filePath == null){
+      throw new IllegalArgumentException("filePath cannot be null.");
+    }
     PrintWriter printWriter = new PrintWriter(new File(filePath));
     printWriter.close();
   }
 
   public void writeLine(String filePath, String line) throws FileNotFoundException {
+    if (filePath == null){
+      throw new IllegalArgumentException("filePath cannot be null.");
+    }
+    if (line == null){
+      throw new IllegalArgumentException("line cannot be null.");
+    }
     PrintWriter printWriter = new PrintWriter(filePath);
     printWriter.println(line);
     printWriter.close();
   }
 
   public List<String> readLines(String filePath) throws FileNotFoundException {
+    if (filePath == null){
+      throw new IllegalArgumentException("filePath cannot be null.");
+    }
     List<String> lines = new ArrayList<>();
     Scanner scanner = new Scanner(new File(filePath));
     while (scanner.hasNextLine()) {
@@ -54,6 +72,9 @@ public class FileHelper {
   }
 
   public String readLastLine(String filePath) throws FileNotFoundException {
+    if (filePath == null){
+      throw new IllegalArgumentException("filePath cannot be null.");
+    }
     Scanner scanner = new Scanner(new File(filePath));
     String lastLine = null;
     while (scanner.hasNextLine()) {
@@ -64,6 +85,9 @@ public class FileHelper {
   }
 
   public void removeLine(int lineNumber) throws FileNotFoundException {
+    if (lineNumber >= 0){
+      throw new IllegalArgumentException("lineNumber cannot less or equal to zero.");
+    }
     List<String> lines = readLines(file.getPath());
     lines.remove(lineNumber - 1);
     PrintWriter printWriter = new PrintWriter(file.getPath());
